# internal/apidocs/gen/swagger.json
{
    "components": {
        "schemas": {
...
            "server.databaseCleanupResponse": {
                "properties": {
                    "cleanup_status": {
                        "$ref": "#/components/schemas/server.databaseCleanupStatus"
                    },
                    "message": {
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "server.databaseCleanupStatus": {
                "properties": {
                    "deleted_tasks": {
                        "type": "integer"
                    },
                    "monitoring_retained": {
                        "type": "string"
                    },
                    "monitoring_storage": {
                        "type": "boolean"
                    },
                    "retention_days": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.databaseFileSizes": {
                "properties": {
                    "database": {
                        "type": "integer"
                    },
                    "shm": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    },
                    "wal": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.databaseInternalStats": {
                "properties": {
                    "freelist_bytes": {
                        "type": "integer"
                    },
                    "freelist_count": {
                        "type": "integer"
                    },
                    "page_count": {
                        "type": "integer"
                    },
                    "page_size": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.databaseStats": {
                "properties": {
                    "databases": {
                        "items": {
                            "$ref": "#/components/schemas/server.databaseStatsDB"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "dialect": {
                        "type": "string"
                    },
                    "total_rows": {
                        "type": "integer"
                    },
                    "total_tables": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "server.databaseStatsDB": {
                "properties": {
                    "files": {
                        "$ref": "#/components/schemas/server.databaseFileSizes"
                    },
                    "internal": {
                        "$ref": "#/components/schemas/server.databaseInternalStats"
                    },
                    "name": {
                        "type": "string"
                    },
                    "storage_path": {
                        "type": "string"
                    },
                    "tables": {
                        "items": {
                            "$ref": "#/components/schemas/server.databaseTableStat"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "total_rows": {
                        "type": "integer"
                    },
                    "total_tables": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.databaseTableStat": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "row_count": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.databaseVacuumResponse": {
                "properties": {
                    "databases": {
                        "items": {
                            "$ref": "#/components/schemas/server.databaseVacuumResult"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "message": {
                        "type": "string"
                    },
                    "size_after": {
                        "type": "integer"
                    },
                    "size_before": {
                        "type": "integer"
                    },
                    "space_reclaimed": {
                        "type": "integer"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "server.databaseVacuumResult": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "size_after": {
                        "type": "integer"
                    },
                    "size_before": {
                        "type": "integer"
                    },
                    "space_reclaimed": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "server.wrappedError": {
                "properties": {
                    "details": {
                        "description": "Optional detail; omitted when empty",
                        "type": "string"
                    },
                    "error": {
                        "description": "Error message",
                        "type": "string"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "timestamp": {
                        "description": "RFC3339 UTC",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "The swag-generated half of the Agent API v1 document. Serve-time authority lives in internal/apidocs: the hand-maintained fragment (openapi.json) owns info/tags/components/security and every not-yet-migrated path, and this generated document's paths merge into it.",
        "title": "Agent API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/database/analyze": {
            "post": {
                "description": "Minimum role: admin. Refreshes query-planner statistics on every open database file; lightweight and safe anytime.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "ANALYZE completed"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.wrappedError"
                                }
                            }
                        },
                        "description": "ANALYZE failed"
                    }
                },
                "summary": "Run ANALYZE",
                "tags": [
                    "Database Management"
                ]
            }
        },
        "/database/cleanup": {
            "post": {
                "description": "Minimum role: admin. The same retention pass the periodic cleanup runs: finished tasks past tasks.retention_days plus stored telemetry past monitoring.retention_days.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.databaseCleanupResponse"
                                }
                            }
                        },
                        "description": "Cleanup triggered"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.wrappedError"
                                }
                            }
                        },
                        "description": "Cleanup failed"
                    }
                },
                "summary": "Trigger manual cleanup",
                "tags": [
                    "Database Management"
                ]
            }
        },
        "/database/stats": {
            "get": {
                "description": "Minimum role: viewer. File sizes (main/WAL/SHM), table row counts, and SQLite internals for EVERY open database file — this agent splits storage by write-contention domain instead of one file.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.databaseStats"
                                }
                            }
                        },
                        "description": "Database statistics"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.wrappedError"
                                }
                            }
                        },
                        "description": "Failed to retrieve database statistics"
                    }
                },
                "summary": "Database statistics",
                "tags": [
                    "Database Management"
                ]
            }
        },
        "/database/vacuum": {
            "post": {
                "description": "Minimum role: admin. Rebuilds every open database file to reclaim space; temporarily doubles disk usage per file.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.databaseVacuumResponse"
                                }
                            }
                        },
                        "description": "VACUUM completed"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/server.wrappedError"
                                }
                            }
                        },
                        "description": "VACUUM failed"
                    }
                },
                "summary": "Run VACUUM",
                "tags": [
                    "Database Management"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}